Refactor publish.py to update the tale structure after publishing and to update the tale yaml file
First 40 lines: I added/removed some imports

create_upload_eml: I renamed the `science_metadata.xml` file to `metadata.xml`. This change was also made in the Dashboard.

Some changes were made to the tale.yaml, which are outlined below.

create_external_object_structure: This structure where we save the md5 is no longer in the spec, so I removed it from the file.

create_metadata_info_structure: This function is responsible for handing the `Metadata` section in the tale yaml.

create_environment_structure: We added the environment section to the spec, this is responsible for filling it out.

create_outer_structure: Fills in the misc details that are outside of any sections in the tale yaml.

create_upload_tale_yaml: Acts as the brain for creating and uploading the tale yaml. I refactored it to use the new functions above.

upload_license_file: Renamed the two variables for the linter.

create_upload_remote_file: In the case where a file exists remotely, we'll want to upload it to DataONE. This method downloads the file and uploads it to DataONE.

publish_tale: First change is that I added progress output to the job_manager. I also added early sanitation, checking if the user included any files in the Tale and that we are able to access the Tale.
Ln 652: This is where we loop over external data and upload them to DataONE.
At the end of the method, the Tale is updated with the package pid, package location, and the published flag is set to True